+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @m5kdev/backend
 
+## 0.20.38
+
+### Patch Changes
+
+- hooks
+- Updated dependencies
+  - @m5kdev/commons@0.20.38
+  - @m5kdev/config@0.20.38
+
+## 0.20.37
+
+### Patch Changes
+
+- hooks
+- Updated dependencies
+  - @m5kdev/commons@0.20.37
+  - @m5kdev/config@0.20.37
+
 ## 0.20.34
 
 ### Patch Changes