+++ packages/email/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @m5kdev/email
 
+## 0.20.38
+
+### Patch Changes
+
+- hooks
+
+## 0.20.37
+
+### Patch Changes
+
+- hooks
+
 ## 0.20.34
 
 ### Patch Changes