+++ packages/web-ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @m5kdev/web-ui
 
+## 0.20.38
+
+### Patch Changes
+
+- hooks
+- Updated dependencies
+  - @m5kdev/frontend@0.20.38
+  - @m5kdev/commons@0.20.38
+
+## 0.20.37
+
+### Patch Changes
+
+- hooks
+- Updated dependencies
+  - @m5kdev/frontend@0.20.37
+  - @m5kdev/commons@0.20.37
+
 ## 0.20.34
 
 ### Patch Changes